Applied one minor patch to get the code to compile on gcc4
git-svn-id: http://svn.musicbrainz.org/libmusicbrainz/trunk@747 b0b80210-5d09-0410-99dd-b4bd03f891c0

+++ lib/comhttpsocket.cpp
@@ -194,7 +194,7 @@ int MBCOMHTTPSocket::NBRead(char* pBuffer, int nLen, int* nBytesWritten, int nTi
 	// advance to the data now, if there is any in this first buffer. 
 	char* pData = strstr(HeaderBuffer, "\r\n\r\n");
 	if (pData) pData += 4;
-	int nOffset = (int)pData - (int)HeaderBuffer;
+	int nOffset = pData - HeaderBuffer;
 	if (nTotal - nOffset >= nLen) // case 1: entire requested read is in header chunk
 	{
 		memcpy(pBuffer, pData, nLen);
